98 add predictive model for records (#99)
* ✨ Add predictive model for records

* 👷 Build project

* ✨ Predict categories

* ✨ Initalize model

+++ api/app/Models/Record.php
@@ -2,15 +2,16 @@
 
 namespace App\Models;
 
-use Illuminate\Database\Eloquent\Model;
-use Database\Factories\RecordFactory;
-use Illuminate\Database\Eloquent\SoftDeletes;
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Factories\Factory;
-use Illuminate\Http\Request;
 use DateTime;
+use Illuminate\Http\Request;
 use App\Services\CurrencyConverter;
+use Database\Factories\RecordFactory;
 use Illuminate\Support\Facades\Cache;
+use App\Http\Controllers\AiController;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 class Record extends Model
 {
@@ -29,12 +30,19 @@ class Record extends Model
 
     protected $hidden = ['category', 'account', 'toAccount', 'import'];
 
+    protected static $disableAiControllerProcessing = false;
+
     public static function boot()
     {
         parent::boot();
 
         static::created(function ($record) {
             Cache::clear();
+
+            if (!static::$disableAiControllerProcessing) {
+                AiController::trainModelWithRecord($record);
+            }
+
             if ($record->type === 'transfer') {
                 static::withoutEvents(function () use ($record) {
                     $transfercategory = 1;
@@ -58,6 +66,7 @@ class Record extends Model
                     ])->save();
                 });
             }
+
         });
 
         static::deleting(function ($record) {
@@ -71,14 +80,19 @@ class Record extends Model
                 });
             }
         });
-
+        
         static::updating(function ($record) {
             Cache::clear();
-            if ($record->type === 'transfer' || $record->original['type'] === 'transfer') {
+
+            if (!static::$disableAiControllerProcessing) {
+                AiController::trainModelWithRecord($record);
+            }
+
+            if ($record->type === 'transfer' || (!empty($record->original['type']) && $record->original['type'] === 'transfer')) {
                 static::withoutEvents(function () use ($record) {
                     $recordAssoc = ($record->link_record_id) ? Record::withTrashed()->find($record->link_record_id) : new Record();
                     if ($recordAssoc) {
-                        if ($record->original['type'] === 'transfer' && $record->type !== 'transfer') {
+                        if (!empty($record->original['type']) && $record->original['type'] === 'transfer' && $record->type !== 'transfer') {
                             $recordAssoc->delete();
                         } else {
                             $transfercategory = 1;
@@ -207,5 +221,15 @@ class Record extends Model
         }
 
         return $query;
+    }
+
+    public static function disableAiControllerProcessing()
+    {
+        static::$disableAiControllerProcessing = true;
+    }
+
+    public static function enableAiControllerProcessing()
+    {
+        static::$disableAiControllerProcessing = false;
     }
 }
